refactored ts_forecast
