+++ scripts/forecast/__init__.py
@@ -1,2 +1,3 @@
 from .interfaces import *  # noqa: F403
 from .ts_forecast import *  # noqa: F403
+from .ts_feature_engineering import *  # noqa: F403
