+++ scripts/forecast/interfaces.py
@@ -135,3 +135,8 @@ class IModel(ABC):
 
         """
         pass
+
+class IFeatureEngineer(ABC):
+    @abstractmethod
+    def transform(self, df, column_name, include_lead):
+        pass
